# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: openleash API
  description: |
    Local-first authorization and cryptographic proof sidecar for AI agents.

    OpenLeash evaluates YAML-based policies against agent action requests and,
    when an action is allowed, issues PASETO v4.public proof tokens that
    counterparties can verify offline or online.

    ## Authentication

    The API uses four security zones:

    - **Public** — No authentication. Health checks, public key distribution,
      proof verification, and the playground.
    - **Agent** — Ed25519 request signing via five headers (`X-Agent-Id`,
      `X-Timestamp`, `X-Nonce`, `X-Body-Sha256`, `X-Signature`). Used for
      `/v1/authorize` and `/v1/agent/*` endpoints.
    - **Owner** — PASETO v4.public session token via `Authorization: Bearer`
      header (or `openleash_session` cookie for GUI). Issued on login.
      Used for all `/v1/owner/*` endpoints.
    - **Admin** — Bearer token and/or localhost check depending on the
      `admin.mode` setting (`localhost`, `token`, or `localhost_or_token`).
      Used for all `/v1/admin/*` endpoints.

    ## Typical flow

    1. An admin creates an **owner** via the Admin API and generates a
       **setup invite**.
    2. The owner completes setup (`POST /v1/owner/setup`) with the invite
       token and a passphrase, then logs in (`POST /v1/owner/login`).
    3. The owner creates **policies** via the Owner API.
    4. An agent calls `/v1/agents/registration-challenge` then
...
servers:
...
paths:
...
  /v1/authorize:
    post:
      summary: Authorize an action
      operationId: authorize
      tags: [Agent]
      description: |
        Main authorization endpoint. The agent submits a signed action
        request and receives a decision (`ALLOW`, `DENY`,
        `REQUIRE_APPROVAL`, `REQUIRE_STEP_UP`, `REQUIRE_DEPOSIT`).

        When the decision is `ALLOW`, a PASETO v4.public proof token is
        included that the agent can present to counterparties.

        When the decision is `REQUIRE_APPROVAL`, the agent should create
        an approval request via `POST /v1/agent/approval-requests`, wait
        for owner approval, then re-authorize with the `approval_token`.

        **Approval token re-authorization**: Include `approval_token` in
        the request body alongside the original action. The server
        verifies the token, checks the action hash matches, marks the
        approval as consumed (single-use), and issues a proof token.

        All five signing headers are required. See the protocol docs
        for the signing input format.
      parameters:
        - name: X-Agent-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Timestamp
          in: header
          required: true
          schema:
            type: string
            format: date-time
        - name: X-Nonce
          in: header
          required: true
          schema:
            type: string
        - name: X-Body-Sha256
          in: header
          required: true
          schema:
            type: string
        - name: X-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ActionRequest"
                - type: object
                  properties:
                    approval_token:
                      type: string
                      description: |
                        Optional PASETO v4.public approval token from an approved
                        HUMAN_APPROVAL request. When present, the server bypasses
                        policy evaluation and directly issues a proof token if the
                        approval token is valid and the action hash matches.
      responses:
        "200":
          description: Authorization decision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizeResponse"
        "400":
          description: Invalid request or approval token error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Agent authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/agent/self:
    get:
      summary: Get own agent info
      operationId: getAgentSelf
      tags: [Agent]
      description: |
        Returns the authenticated agent's own record. Requires Ed25519
        request signing.
      parameters:
        - name: X-Agent-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Timestamp
          in: header
          required: true
          schema:
            type: string
            format: date-time
        - name: X-Nonce
          in: header
          required: true
          schema:
            type: string
        - name: X-Body-Sha256
          in: header
          required: true
          schema:
            type: string
        - name: X-Signature
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent info.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRegistration"
  /v1/agent/approval-requests:
    post:
      summary: Create approval request
      operationId: createApprovalRequest
      tags: [Agent]
      description: |
        Creates a human approval request for an action that received
        a `REQUIRE_APPROVAL` decision. The request is stored as PENDING
        and the owner can approve or deny it via the Owner API.

        Requires Ed25519 request signing.
      parameters:
        - name: X-Agent-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Timestamp
          in: header
          required: true
          schema:
            type: string
            format: date-time
        - name: X-Nonce
          in: header
          required: true
          schema:
            type: string
        - name: X-Body-Sha256
          in: header
          required: true
          schema:
            type: string
        - name: X-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision_id, action]
              properties:
                decision_id:
                  type: string
                  format: uuid
                action:
                  $ref: "#/components/schemas/ActionRequest"
                justification:
                  type: string
                  description: Human-readable reason for the action.
                context:
                  type: object
                  description: Additional context for the owner's review.
      responses:
        "200":
          description: Approval request created.
          content:
            application/json:
              schema:
                type: object
                required: [approval_request_id, status, expires_at]
                properties:
                  approval_request_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [PENDING]
                  expires_at:
                    type: string
                    format: date-time
  /v1/agent/approval-requests/{approvalRequestId}:
    get:
      summary: Get approval request status
      operationId: getAgentApprovalRequest
      tags: [Agent]
      description: |
        Returns the status of an approval request. If the request has
        been approved, the response includes the `approval_token` that
        the agent should include in its re-authorization request.

        Scoped to the authenticated agent's own requests.
        Requires Ed25519 request signing.
      parameters:
        - $ref: "#/components/parameters/ApprovalRequestId"
        - name: X-Agent-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Timestamp
          in: header
          required: true
          schema:
            type: string
            format: date-time
        - name: X-Nonce
          in: header
          required: true
          schema:
            type: string
        - name: X-Body-Sha256
          in: header
          required: true
          schema:
            type: string
        - name: X-Signature
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approval request details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequestAgentView"

  # ── Owner ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Admin token configured in `config.yaml` under `admin.token`.
        Required when `admin.mode` is `token` or for remote access in
        `localhost_or_token` mode.

    OwnerSession:
      type: http
      scheme: bearer
      description: |
        PASETO v4.public session token issued by `POST /v1/owner/login`.
        Contains `purpose: 'owner_session'` claim to distinguish from
        proof tokens. Default TTL is 8 hours (configurable via
        `sessions.ttl_seconds` in `config.yaml`).
  parameters:
...
    ApprovalRequestId:
      name: approvalRequestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the approval request.
  schemas:
    Error:
      type: object
      description: Standard error envelope returned by all endpoints on failure.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code (e.g. `INVALID_REQUEST`, `AGENT_NOT_FOUND`).
            message:
              type: string
              description: Human-readable error message.
            details:
              type: object
              description: Additional structured details (e.g. Zod validation errors).
...
    ActionRequest:
      type: object
      description: |
        An action the agent wants to perform. This is the input to the
        authorization engine which evaluates it against the bound policy.
      required: [action_type, principal, payload]
      properties:
        action_id:
          type: string
          format: uuid
          description: Unique ID for this action. Auto-generated if omitted.
        action_type:
          type: string
          description: The type of action (e.g. `transfer`, `purchase`, `sign`). Matched against policy rule patterns.
        requested_at:
          type: string
          format: date-time
          description: When the action was requested. Defaults to current time.
        principal:
          type: object
          required: [agent_id]
          description: The agent requesting the action.
          properties:
            agent_id:
              type: string
              description: The agent's registered `agent_id`.
        subject:
          type: object
          description: The owner principal on whose behalf the agent is acting.
          properties:
            principal_id:
              type: string
              format: uuid
        resource:
          type: object
          description: The resource being acted upon.
          properties:
            resource_id:
              type: string
            resource_type:
              type: string
        relying_party:
          type: object
          description: The counterparty that will receive the proof token.
          properties:
            rp_id:
              type: string
              format: uuid
            domain:
              type: string
            trust_profile:
              type: string
              enum: [LOW, MEDIUM, HIGH, REGULATED]
              description: Trust tier of the counterparty. Included in proof token claims.
        payload:
          type: object
          description: Action-specific data evaluated by policy constraints (e.g. `amount`, `currency`).
        context:
          type: object
          description: Additional context available to policy expressions (e.g. IP address, device info).
    Obligation:
      type: object
      description: |
        A condition that must be fulfilled before the action can proceed.
        Returned when the decision is `REQUIRE_APPROVAL`, `REQUIRE_STEP_UP`,
        or `REQUIRE_DEPOSIT`.
      required: [obligation_id, type, status]
      properties:
        obligation_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [HUMAN_APPROVAL, STEP_UP_AUTH, DEPOSIT, COUNTERPARTY_ATTESTATION]
          description: The kind of obligation that must be fulfilled.
        status:
          type: string
          enum: [PENDING, FULFILLED, WAIVED]
        details_json:
          type: object
          description: Obligation-specific details (e.g. required deposit amount).
    AuthorizeResponse:
      type: object
      description: |
        The authorization decision. When `result` is `ALLOW`, a PASETO
        proof token is included that the agent can present to counterparties.
      required: [decision_id, action_id, action_hash, result, matched_rule_id, reason, proof_token, proof_expires_at, obligations]
      properties:
        decision_id:
          type: string
          format: uuid
          description: Unique ID for this decision (recorded in audit log).
        action_id:
          type: string
          format: uuid
        action_hash:
          type: string
          description: SHA-256 hash of the canonicalized action. Used for proof verification.
        result:
          type: string
          enum: [ALLOW, DENY, REQUIRE_APPROVAL, REQUIRE_STEP_UP, REQUIRE_DEPOSIT]
          description: The authorization decision.
        matched_rule_id:
          type: string
          nullable: true
          description: ID of the policy rule that produced this decision, or null if no rule matched.
        reason:
          type: string
          description: Human-readable explanation of the decision.
        proof_token:
          type: string
          nullable: true
          description: PASETO v4.public token. Present only when `result` is `ALLOW`.
        proof_expires_at:
          type: string
          format: date-time
          nullable: true
          description: Expiration time of the proof token.
        obligations:
          type: array
          items:
            $ref: "#/components/schemas/Obligation"
          description: Obligations that must be fulfilled (empty array when `result` is `ALLOW` or `DENY`).
    AgentRegistration:
      type: object
      description: A registered AI agent.
      required: [agent_principal_id, agent_id, owner_principal_id, status, created_at]
      properties:
        agent_principal_id:
          type: string
          format: uuid
          description: Server-assigned UUID for this agent.
        agent_id:
          type: string
          description: The agent's self-chosen identifier.
        owner_principal_id:
          type: string
          format: uuid
          description: UUID of the owner this agent belongs to.
        status:
          type: string
          description: Agent status (e.g. `ACTIVE`, `REVOKED`).
        created_at:
          type: string
          format: date-time
...
    ApprovalRequestAgentView:
      type: object
      description: Agent-facing view of an approval request.
      required: [approval_request_id, status, action_type, action_hash, created_at, expires_at]
      properties:
        approval_request_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, APPROVED, DENIED, EXPIRED]
        action_type:
          type: string
        action_hash:
          type: string
        justification:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true
        denial_reason:
          type: string
          nullable: true
        approval_token:
          type: string
          description: Present only when status is APPROVED.
        approval_token_expires_at:
          type: string
          format: date-time
          description: Present only when status is APPROVED.
...